Deprecates get_userinfo() master function.
get_userinfo() function has a very ill-defined interface.
It returns an array which is usually accessed by the
*internal* symbolic constants (for indexing the
*internal* userlist structure) from userinfo.h plus an
offset. This is error-prone and the internal indices
can therefore never be changed internally.
It is also completely undocumented.

The alternative is query_userlist(). If a lot of
people need all data from the userlist at once,
an additional query function should be designed.

Change-Id: I848b764f46f609f612d1847c1e0198cd3dc454ab